requirements.txt
fastapi==0.111.0
google_api_python_client==2.133.0
google_auth_oauthlib==1.2.0
passlib==1.7.4
pydantic==2.7.4
pymongo==4.6.3
python-dotenv==1.0.1
python_jose==3.3.0
Requests==2.32.3
bcrypt==4.2.0
uvicorn==0.30.3